Allow replacement of bad encoding
When processing CSV files we may encounter fields which are
poorly encoded. These fields may not be of concern to the
task at hand and such bad encoding can be ignored.

Herein an option `:replacer` is introduced, the value of which
is used to replace bad codepoints. It defaults to `nil` which disables
any replacement.